requirements.txt
aiogoogletrans==3.3.2
aiohttp==3.5.4
altgraph==0.16.1
astroid==2.2.5
async-timeout==3.0.1
asyncurban==0.3.3
attrs==19.1.0
beautifulsoup4==4.8.0
bitlyshortener==0.3.2
braceexpand==0.1.5
cachetools==3.1.1
certifi==2019.9.11
chardet==3.0.4
Click==7.0
cowpy==1.1.0
dadjokes==1.3
dblpy==0.3.3
discord==1.0.1
discord.py==1.2.5
dnspython==1.15.0
dnspython3==1.15.0
forex-python==1.5
fortune==1.1.0
geojson==2.5.0
grizzled-python==2.2.0
hastebin.py==0.2
humanize==0.5.1
idna==2.8
import-expression==0.7.0
ipinfo==2.0.0
isort==4.3.21
jishaku==1.17.0.169
lavalink==3.0.0
lazy-object-proxy==1.4.1
mccabe==0.6.1
mcstatus==2.2.1
mtranslate==1.6
multidict==4.5.2
pbr==5.4.3
psutil==5.6.3
pycosat==0.6.3
PyHamcrest==1.9.0
pylint==2.3.1
pynput==1.4.2
pyowm==2.10.0
python-xlib==0.25
ratelimiter==1.2.0.post0
requests==2.22.0
ruamel.yaml==0.16.5
ruamel.yaml.clib==0.1.2
simplejson==3.16.0
six==1.12.0
soupsieve==1.9.4
StringGenerator==0.3.3
testresources==2.0.1
timeago==1.0.10
typed-ast==1.4.0
urllib3==1.25.6
websockets==6.0
wrapt==1.11.2
xmltodict==0.12.0
yarl==1.3.0
